kgserver/pyproject.toml
[project]
name = "kgserver"
version = "0.1.0"
description = "Containerized server for knowledge graph data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kgbundle",
    "black[jupyter]>=26.1.0",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.4",
    "flake8>=7.3.0",
    "httpx>=0.27.0",
    "mcp>=1.25.0",
    "mypy>=1.19.1",
    "psycopg-binary>=3.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruamel-yaml>=0.19.1",
    "ruff>=0.14.13",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "strawberry-graphql[fastapi]>=0.289.2",
    "typing>=3.10.0.0",
    "uvicorn>=0.40.0",
]

[tool.black]
line-length = 200

[tool.ruff]
line-length = 200

[tool.ruff.lint]
ignore = ["E402"]

[tool.pylint.main]
max-line-length = 200

[tool.flake8]
line-length = 200
ignore = ["E203", "E266", "E501", "W503", "E402"]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
packages = ["storage", "query", "mcp_server"]


